sua vai thu hay ho

+++ authes/tbKH.js
@@ -12,19 +12,9 @@ function createRow(rowData) {
           <td>${rowData.quoctich}</td>
           <td>${rowData.cccd}</td>
           <td>${rowData.hochieu}</td>
-          <td><a href="form-elementpt.html?id=${
-            rowData.id
-          }"><button class="btn btn-danger btn-anim" onclick="updateKHById(${
-    rowData.id
-  });"><i class="fa fa-pencil-square-o"></i><span class="btn-text">Booking</span></button></a></td>  
-          <td><a href="form-editkh.html?id=${
-            rowData.id
-          }"><button class="btn btn-danger btn-anim" onclick="updateKHById(${
-    rowData.id
-  });"><i class="fa fa-pencil-square-o"></i><span class="btn-text">edit</span></button></a></td>       
-          <td><button class="btn btn-danger btn-anim" onclick="delKH(${
-            rowData.id
-          })"><i class="fa fa-trash-o"></i><span class="btn-text">delete</span></button></td>
+          
+          <td><a href="form-editkh.html?id=${rowData.id}"><button class="btn btn-info btn-icon-anim btn-square" onclick="updateKHById(${rowData.id});"><i class="fa fa-pencil-square-o"></i></button></a></td>
+          <td><button class="btn btn-info btn-icon-anim btn-square" onclick="delKH(${rowData.id})"><i class="icon-trash"></i></button></td>
       </tr>
     `;
   return row;
@@ -43,6 +33,8 @@ function createRow(rowData) {
 })();
 
 function delKH(id) {
-  deleteKH(id);
-  window.location.reload();
+  if (confirm("Do you want Delete?")) {
+    deleteKH(id);
+    window.location.reload();
+  }
 }
