+++ authes/tbNV.js
@@ -9,13 +9,9 @@ function createRow(rowData) {
           <td>${getDate(rowData.dob)}</td>
           <td>${rowData.phone}</td>
           <td>${rowData.gender}</td>
-          <td><a href="form-editnv.html?id=${rowData.id}">
-          <button class="btn btn-danger btn-anim" onclick="updateNVById(${
-            rowData.id
-          });"><i class="fa fa-pencil-square-o"></i><span class="btn-text">edit</span></button></a></td>       
-          <td><button class="btn btn-danger btn-anim" onclick="delNV(${
-            rowData.id
-          })"><i class="fa fa-trash-o"></i><span class="btn-text">delete</span></button></td>
+        
+          <td><a href="form-editnv.html?id=${rowData.id}"><button class="btn btn-info btn-icon-anim btn-square" onclick="updateNVById(${rowData.id});"><i class="fa fa-pencil-square-o"></i></button></a></td>
+          <td><button class="btn btn-info btn-icon-anim btn-square" onclick="delNV(${rowData.id})"><i class="icon-trash"></i></button></td>
       </tr>
     `;
   return row;
@@ -34,6 +30,8 @@ function createRow(rowData) {
 })();
 
 function delNV(id) {
-  deleteNV(id);
-  window.location.reload();
+  if (confirm("Do you want Delete?")) {
+    deleteNV(id);
+    window.location.reload();
+  }
 }
